refactor remaining mesh setup style controls to be ParticleStyle
